requirements.txt
django==4.2.15
django-filter==24.3